Revert "libseat: Check euid before using builtin"
This reverts commit 1ae6c3b3ddf0ce2a2e1817eb9c74e0c03153df58.

A user might want to run the builtin server as non root, if they have
permission to use the devices.

The check was originally copied from wlroots's direct backend. It was reverted
in fa05d3cde68d with a detailed explanation of why root priviledges are not
always necessary to use the DRM device.