seat: WIP grace period
When a display server terminates, seatd resets the VT to a safe state
with the kernel console activated, even if another display server will
start soon after that. This forces blanking.

Instead, allow for a small grace period, during which a new display
server may start without VT resets. If the grace period expires, the VT
resets as normal. A very short delay should be sufficient.