logind: Remove drm device tracking
Tying the session state to active drm devices is fragile and does not
make sense. Instead, rely mainly on PropertiesChanged for the session
state. The session is also suspended on the first gentle device pause as
these must be acked.

This also removes the need for listening to device resume, as the
PropertiesChanged handler takes care of re-activation.