libseat/seatd: Read remaining events after processing responses
When sending a request and awaiting a response, the client will read
data into the ringbuffer and queue events until the response is found.
Normally we would not have received any events, but if we did, we issue
a ping request to seatd so that the libseat caller will be woken up for
a dedicated dispatch.

If an event is received immediately after the response such that we will
read both into ringbuffer at the same time, but we stop reading the
ringubffer to queue events once we hit the response. As such, the event
will not be queued and we will not attempt a ping to ensure an explicit
dispatch occurs. At this point we are deadlocked, as the libseat caller
is unaware that an explicit dispatch is needed to process events already
read from the socket.

Before checking if any events are queued, read any remaining events and
queue them so that we do not leave full messages in the ring buffer.
This should fix an issue where when switching session, the libseat
client sometimes fails to see the disable_seat event.