logind: Use a constant for the "Active" field when log
In the property change handler, the variable `s` is read using:

    ret = sd_bus_message_read_basic(msg, 's', &s);

However, after further calls like:

    ret = sd_bus_message_read_basic(msg, 'b', &value);

the contents of `s` may become invalid or corrupted. This is because
string-like values returned by sd_bus_message_read_basic() point to
internal memory owned by the message and are only valid until the
next read or until the message is released.

To ensure correctness, this patch copies the string into a local
buffer immediately after reading. This avoids referencing stale or
invalid memory when logging or comparing the property key.

Reference:
https://www.freedesktop.org/software/systemd/man/sd_bus_message_read_basic.html

Signed-off-by: Hongfei Shang <shanghongfei@kylinos.cn>
Signed-off-by: Hongfei Shang <shanghongfei@zohomail.com>